Hypothetical fix for the public utility commands not appearing when the bot joins a server, and instead only appearing when restarted. (#146)
