+++ src/main/kotlin/net/irisshaders/lilybot/LilyBot.kt
@@ -25,6 +25,7 @@ import net.irisshaders.lilybot.extensions.util.Github
 import net.irisshaders.lilybot.extensions.util.ModUtilities
 import net.irisshaders.lilybot.extensions.util.PublicUtilities
 import net.irisshaders.lilybot.extensions.util.RoleMenu
+import net.irisshaders.lilybot.extensions.util.StartupHooks
 import net.irisshaders.lilybot.extensions.util.Tags
 import net.irisshaders.lilybot.extensions.util.ThreadControl
 import net.irisshaders.lilybot.utils.BOT_TOKEN
@@ -75,6 +76,7 @@ suspend fun main() {
 			add(::PublicUtilities)
 			add(::Report)
 			add(::RoleMenu)
+			add(::StartupHooks)
 			add(::Tags)
 			add(::TemporaryModeration)
 			add(::TerminalModeration)
